refactor: 시간 측정 출력 결과 실수형으로 변경

+++ src/main/java/com/endlessmilkyway/tasks/task01/view/OutputView.java
@@ -27,7 +27,7 @@ public class OutputView {
         printArrayElems(arr);
     }
 
-    public void printSortingProcessTime(long diffTime) {
+    public void printSortingProcessTime(double diffTime) {
         System.out.println(PROCESS_TIME + diffTime);
     }
 }
